Version of sbt changed to 0.13.13
The commit is related to issue #33 (Dockerfile.test error). The newest version of sbt 0.13.15 released on 2017-04-09 depends on openjdk-8-jdk, but openjdk-8-jdk is not available in default apt repos for debian:jessie. Version 0.13.13 (that was the latest in the repo till ~2017-04-09) doesn't depends on openjdk-8-jdk and doesn't cause problem with building the docker image.